Add k/d bar pages.
This is similar to the horizontal killboard from kqstats, with the addition of assists. The goal is to add additional stats, likely based on role. E.g. berry drones show runs, kicks, assists, and deaths.

The template is split into pieces to accommodate a future multi-view page. For example, if a page included both blue and gold statsboards, the CSS and JS sections would be included on the page once; the JS_init and Body would be included once with Side=blue and once with Side=gold; either there would need to be a way to specify positioning of each via URL, and have that affect the source, or rely on the user to add appropriate layout CSS.