Hooks game tracker changes back to websocket outputs in the "control" section.
The score controller page can now receive external changes. If there are two such pages open, changes in one will appear on the other.
Currently this assumes that team A is on blue, which is fine until something gets hooked up to the swap-sides function.
Changes to the teams or scores currently send both, there aren't checks if an update actually changed anything before sending, and the controller always triggers all three updates together, so any change on the scoreboard will result in three packets coming from the server.