Add a scoreboard page, which shows current teams and scores.
The scoreboard is configured via configuration template, which overrides templates defined in the scoreboard.tpl file. Sometime in the future I'll add support for a clearer config file for just configuring the default components, since it's fairly difficult right now but that should be common.
This is based on the overlay from https://github.com/ughoavgfhw/kqstats, but without a background. The background can be displayed directly in OBS so it doesn't need to be transmitted and processed as a web page.

Added a comment where I realized there is a race condition between changing match scores and sending messages about previous changes. The last change will always happen before the last sent message, so this should be relatively benign, but I'll still get rid of the issue in a future redesign.