Auto-increment scores when a game ends. Currently does not finish or advance matches.
Ideally this could be turned on and off by some option, but for now it is always on.
I also want to try having a button to send the `mapselect` command to the cab, which if it works would cancel the active game, therefore preventing the score from being updated. This could be used for warmup and button testing games.