Add the team pictures page.
Parses player names from the teams file.
Adds a view which shows the two current teams, using the existing currentMatch section. A new tournamentData section provides the mapping from team names to players and photo URIs.
Photos are read from a photos/ directory by attempting to open the player's name with a few different extensions. If none work but a lookup of "default" finds a match, that is used.