Add support for a JSON config file
Upcoming features will require more configuration than one would want to specify on the command line.
The existing flags/command line arguments will still work and override the config file.
The default printing of a model to the command line is removed as a result. Explicitly specifying a model will add it back.