Start cleanup of data movement between sections of the binary
All data is collected in an Event, which is currently but need not be triggered by a cab message. This allows both internally generated events, and processing all data generated by a message in one go for sending announcements.
Currently the event is created and passed to the server, which then expands it to the old format.