Move the last uses of the update channel in the tracker to events.
Advancing a match and switching sides both add team and score updates. Advancing also adds a victory rule update.
Setting on-deck teams never sent updates, and since it needs to be more generic, doing so now is not useful.
None of these commands is actually hooked to anything right now.